Adding Problem instances

+++ NumberPartition/src/main/java/edu/harvard/cscie124/pa3/NumberPartitionSolver.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface NumberPartitionSolver {
 
-	public int getResidue(List<Integer> list);
+	public long getResidue(List<Long> list);
 	
 }
